requirements.txt
web3==6.15.1
eth-abi>=5.2.0
python-dotenv>=1.2.1
aiohttp>=3.13.2
requests>=2.32.5
psycopg2-binary>=2.9.11
simplejson>=3.20.2
colorama>=0.4.6
websockets>=15.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
